updated to get source files from installed version

+++ src/fastlisaresponse/cutils/LISAResponse.cu
@@ -236,7 +236,7 @@ void interp(double *result_hp, double *result_hc, cmplx *input, int h, int d, do
 #define MAX_ORDER 40
 
 CUDA_KERNEL
-void TDI_delay(double *delayed_links, double *input_links, int num_inputs, int num_delays, double *t_arr, int *unit_starts, int *unit_lengths, int *tdi_base_link, int *tdi_link_combinations, int *tdi_signs_in, int *channels, int num_units, int num_channels,
+void TDI_delay(double *delayed_links, double *input_links, int num_inputs, int num_delays, double *t_arr, int *unit_starts, int *unit_lengths, int *tdi_base_link, int *tdi_link_combinations, double *tdi_signs_in, int *channels, int num_units, int num_channels,
                int order, double sampling_frequency, int buffer_integer, double *A_in, double deps, int num_A, double *E_in, int tdi_start_ind, Orbits *orbits_in)
 {
     Orbits orbits = *orbits_in;
@@ -327,7 +327,7 @@ void TDI_delay(double *delayed_links, double *input_links, int num_inputs, int n
             int base_link = tdi_base_link[unit_i];
             int base_link_index = orbits.get_link_ind(base_link);
             int channel = channels[unit_i];
-            int sign = tdi_signs_in[unit_i];
+            double sign = tdi_signs_in[unit_i];
             delay = t;
             for (int sub_i = 0; sub_i < unit_length; sub_i += 1)
             {
@@ -412,7 +412,7 @@ void TDI_delay(double *delayed_links, double *input_links, int num_inputs, int n
     }
 }
 
-void get_tdi_delays(double *delayed_links, double *input_links, int num_inputs, int num_delays, double *t_arr, int *unit_starts, int *unit_lengths, int *tdi_base_link, int *tdi_link_combinations, int *tdi_signs_in, int *channels, int num_units, int num_channels,
+void get_tdi_delays(double *delayed_links, double *input_links, int num_inputs, int num_delays, double *t_arr, int *unit_starts, int *unit_lengths, int *tdi_base_link, int *tdi_link_combinations, double *tdi_signs_in, int *channels, int num_units, int num_channels,
                     int order, double sampling_frequency, int buffer_integer, double *A_in, double deps, int num_A, double *E_in, int tdi_start_ind, Orbits *orbits_in)
 {
