+++ src/fastlisaresponse/cutils/LISAResponse.hh
@@ -16,7 +16,7 @@ void get_response(double *y_gw, double *t_data, double *k_in, double *u_in, doub
                   int projections_start_ind,
                   Orbits *orbits);
 
-void get_tdi_delays(double *delayed_links, double *input_links, int num_inputs, int num_delays, double *t_arr, int *unit_starts, int *unit_lengths, int *tdi_base_link, int *tdi_link_combinations, int *tdi_signs_in, int *channels, int num_units, int num_channels,
+void get_tdi_delays(double *delayed_links, double *input_links, int num_inputs, int num_delays, double *t_arr, int *unit_starts, int *unit_lengths, int *tdi_base_link, int *tdi_link_combinations, double *tdi_signs_in, int *channels, int num_units, int num_channels,
                     int order, double sampling_frequency, int buffer_integer, double *A_in, double deps, int num_A, double *E_in, int tdi_start_ind, Orbits *orbits);
 
 #endif // __LISA_RESPONSE__
